Login Authentication  & UI

+++ database/migrations/A_2014_10_12_000000_create_users_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('prefix_id')->default(1);
+            $table->integer('prefix_id')->default(1);
             $table->string('name');
             $table->string('email')->unique();
-            $table->unsignedBigInteger('country_id')->default(62);
+            $table->integer('country_id')->default(62);
             $table->string('phone');
             $table->string('postal');
             $table->string('password');
